+++ src/components/EditSections/EditUserInfo/EditUserInfo.js
@@ -1,11 +1,12 @@
 import _ from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import { change, Field, formValueSelector, getFormValues } from 'redux-form';
+import { Field } from 'react-final-form';
 import { FormattedMessage, injectIntl } from 'react-intl';
 import moment from 'moment-timezone';
+import { OnChange } from 'react-final-form-listeners';
 
+import SafeHTMLMessage from '@folio/react-intl-safe-html';
 import { ViewMetaData } from '@folio/stripes/smart-components';
 import {
   Button,
@@ -20,6 +21,8 @@ import {
   ModalFooter,
 } from '@folio/stripes/components';
 
+import asyncValidateField from '../../validators/asyncValidateField';
+
 import css from './EditUserInfo.css';
 
 class EditUserInfo extends React.Component {
@@ -30,28 +33,24 @@ class EditUserInfo extends React.Component {
     intl: PropTypes.object.isRequired,
     onToggle: PropTypes.func,
     patronGroups: PropTypes.arrayOf(PropTypes.object),
-    selectedPatronGroup: PropTypes.string,
     stripes: PropTypes.shape({
       store: PropTypes.shape({
         dispatch: PropTypes.func.isRequired,
         getState: PropTypes.func,
       }),
     }).isRequired,
+    form: PropTypes.object,
+    uniquenessValidator: PropTypes.object,
   };
 
   constructor(props) {
     super(props);
 
+    const { initialValues: { patronGroup } } = props;
     this.state = {
       showRecalculateModal: false,
+      selectedPatronGroup: patronGroup,
     };
-  }
-
-  componentDidUpdate(prevProps) {
-    const offsetOfSelectedPatronGroup = this.getPatronGroupOffset();
-    if (this.props.selectedPatronGroup !== prevProps.selectedPatronGroup && offsetOfSelectedPatronGroup !== '') {
-      this.showModal(true);
-    }
   }
 
   showModal = (val) => {
@@ -64,15 +63,31 @@ class EditUserInfo extends React.Component {
     this.setState({ showRecalculateModal: false });
   }
 
-  recalculateExpirationDate = () => {
-    const offsetOfSelectedPatronGroup = this.props.selectedPatronGroup ? this.getPatronGroupOffset() : '';
-    const recalculatedDate = (moment().add(offsetOfSelectedPatronGroup, 'd').format('YYYY-MM-DD'));
-    this.props.stripes.store.dispatch(change('userForm', 'expirationDate', recalculatedDate));
+  setRecalculatedExpirationDate = () => {
+    const { form: { change } } = this.props;
+    const recalculatedDate = this.calculateNewExpirationDate();
+
+    change('expirationDate', recalculatedDate);
     this.setState({ showRecalculateModal: false });
   }
 
+  calculateNewExpirationDate = () => {
+    const { initialValues } = this.props;
+    const expirationDate = new Date(initialValues.expirationDate);
+    const now = Date.now();
+    const offsetOfSelectedPatronGroup = this.state.selectedPatronGroup ? this.getPatronGroupOffset() : '';
+    let recalculatedDate;
+
+    if (initialValues.expirationDate === undefined || expirationDate <= now) {
+      recalculatedDate = (moment().add(offsetOfSelectedPatronGroup, 'd').format('YYYY-MM-DD'));
+    } else {
+      recalculatedDate = (moment(expirationDate).add(offsetOfSelectedPatronGroup, 'd').format('YYYY-MM-DD'));
+    }
+    return recalculatedDate;
+  }
+
   getPatronGroupOffset = () => {
-    const selectedPatronGroup = this.props.patronGroups.find(i => i.id === this.props.selectedPatronGroup);
+    const selectedPatronGroup = this.props.patronGroups.find(i => i.id === this.state.selectedPatronGroup);
     return _.get(selectedPatronGroup, 'expirationOffsetInDays', '');
   };
 
@@ -84,7 +99,10 @@ class EditUserInfo extends React.Component {
       onToggle,
       accordionId,
       intl,
+      uniquenessValidator,
     } = this.props;
+
+    const { barcode } = initialValues;
 
     const isUserExpired = () => {
       const expirationDate = new Date(initialValues.expirationDate);
@@ -93,10 +111,9 @@ class EditUserInfo extends React.Component {
     };
 
     const willUserExtend = () => {
-      const { stripes: { store } } = this.props;
-      const state = store.getState();
-      const formValues = getFormValues('userForm')(state) || {};
-      const currentExpirationDate = new Date(_.get(formValues, 'expirationDate', ''));
+      const { form } = this.props;
+      const expirationDate = form.getFieldState('expirationDate')?.value ?? '';
+      const currentExpirationDate = new Date(expirationDate);
       const now = Date.now();
       return currentExpirationDate >= now;
     };
@@ -108,7 +125,7 @@ class EditUserInfo extends React.Component {
     };
 
     const checkShowRecalculateButton = () => {
-      const offsetOfSelectedPatronGroup = this.props.selectedPatronGroup ? this.getPatronGroupOffset() : '';
+      const offsetOfSelectedPatronGroup = this.state.selectedPatronGroup ? this.getPatronGroupOffset() : '';
       return offsetOfSelectedPatronGroup !== '';
     };
 
@@ -136,14 +153,16 @@ class EditUserInfo extends React.Component {
     ];
 
     const offset = this.getPatronGroupOffset();
-    const group = _.get(this.props.patronGroups.find(i => i.id === this.props.selectedPatronGroup), 'group', '');
+    const group = _.get(this.props.patronGroups.find(i => i.id === this.state.selectedPatronGroup), 'group', '');
+    const date = moment(this.calculateNewExpirationDate()).format('LL');
+
     const modalFooter = (
       <ModalFooter>
         <Button
           id="expirationDate-modal-recalculate-btn"
-          onClick={this.recalculateExpirationDate}
+          onClick={this.setRecalculatedExpirationDate}
         >
-          <FormattedMessage id="ui-users.information.recalculate.expirationDate" />
+          <FormattedMessage id="ui-users.information.recalculate.modal.button" />
         </Button>
         <Button
           id="expirationDate-modal-cancel-btn"
@@ -220,6 +239,15 @@ class EditUserInfo extends React.Component {
                 aria-required="true"
                 required
               />
+              <OnChange name="patronGroup">
+                {(selectedPatronGroup) => {
+                  this.setState({ selectedPatronGroup }, () => {
+                    if (this.getPatronGroupOffset()) {
+                      this.showModal(true);
+                    }
+                  });
+                }}
+              </OnChange>
             </Col>
             <Col xs={12} md={3}>
               <Field
@@ -257,7 +285,7 @@ class EditUserInfo extends React.Component {
               {checkShowRecalculateButton() && (
                 <Button
                   id="recalculate-expirationDate-btn"
-                  onClick={this.recalculateExpirationDate}
+                  onClick={this.setRecalculatedExpirationDate}
                 >
                   <FormattedMessage id="ui-users.information.recalculate.expirationDate" />
                 </Button>
@@ -269,6 +297,7 @@ class EditUserInfo extends React.Component {
                 name="barcode"
                 id="adduser_barcode"
                 component={TextField}
+                validate={asyncValidateField('barcode', barcode, uniquenessValidator)}
                 fullWidth
               />
             </Col>
@@ -281,9 +310,9 @@ class EditUserInfo extends React.Component {
           open={this.state.showRecalculateModal}
         >
           <div>
-            <FormattedMessage
+            <SafeHTMLMessage
               id="ui-users.information.recalculate.modal.text"
-              values={{ group, offset }}
+              values={{ group, offset, date }}
             />
           </div>
         </Modal>
@@ -292,10 +321,4 @@ class EditUserInfo extends React.Component {
   }
 }
 
-const selectFormValue = formValueSelector('userForm');
-
-export default connect(
-  store => ({
-    selectedPatronGroup: selectFormValue(store, 'patronGroup'),
-  })
-)(injectIntl(EditUserInfo));
+export default injectIntl(EditUserInfo);