+++ src/components/PatronBlock/PatronBlockModal.js
@@ -2,16 +2,26 @@ import { take, orderBy } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
+
+import {
+  IfPermission,
+} from '@folio/stripes/core';
 import {
   Button,
   Col,
   Modal,
-  Row
+  Row,
 } from '@folio/stripes/components';
 
 import css from './PatronBlockModal.css';
 
-const PatronBlockModal = ({ open, onClose, patronBlocks, viewUserPath }) => {
+const PatronBlockModal = ({
+  open,
+  onOverride,
+  onClose,
+  patronBlocks,
+  viewUserPath,
+}) => {
   const blocks = take(orderBy(patronBlocks, ['metadata.updatedDate'], ['desc']), 3);
   const renderBlocks = blocks.map(block => {
     return (
@@ -39,16 +49,26 @@ const PatronBlockModal = ({ open, onClose, patronBlocks, viewUserPath }) => {
       <Row>
         <Col xs>
           <FormattedMessage id="ui-users.blocks.reason" />
-          :
         </Col>
       </Row>
       {renderBlocks}
       <br />
       <Row>
-        <Col xs={8}>{(patronBlocks.length > 3) && <FormattedMessage id="ui-users.blocks.additionalReasons" />}</Col>
-        <Col xs={4}>
+        <Col xs={6}>{(patronBlocks.length > 3) && <FormattedMessage id="ui-users.blocks.additionalReasons" />}</Col>
+        <Col xs={6}>
           <Row end="xs">
             <Col>
+              <IfPermission perm="ui-users.overridePatronBlock">
+                <Button
+                  data-test-open-override-modal
+                  onClick={() => {
+                    onClose();
+                    onOverride();
+                  }}
+                >
+                  <FormattedMessage id="ui-users.blocks.overrideButton" />
+                </Button>
+              </IfPermission>
               <Button
                 id="patron-block-close-modal"
                 onClick={onClose}
@@ -73,6 +93,7 @@ const PatronBlockModal = ({ open, onClose, patronBlocks, viewUserPath }) => {
 
 PatronBlockModal.propTypes = {
   open: PropTypes.bool,
+  onOverride: PropTypes.func.isRequired,
   onClose: PropTypes.func,
   patronBlocks: PropTypes.arrayOf(PropTypes.object),
   viewUserPath: PropTypes.string,