+++ src/components/util/util.js
@@ -137,3 +137,7 @@ export function hasEveryLoanItemStatus(loans, itemStatus) {
 export function hasAnyLoanItemStatus(loans, itemStatuses) {
   return _.every(Object.values(loans), loan => itemStatuses.includes(loan?.item?.status?.name));
 }
+
+export function accountsMatchStatus(accounts, status) {
+  return accounts.every((account) => account.status.name.toLowerCase() === status.toLowerCase());
+}