+++ test/bigtest/network/config.js
@@ -271,9 +271,8 @@ export default function config() {
     return this.serializerOrRegistry.serialize(loans.all());
   });
 
-  this.get('loan-storage/loans/:loanid', {
-    loans: [],
-    totalRecords: 0
+  this.get('/loan-storage/loans/:id', ({ loans }, request) => {
+    return loans.find(request.params.id).attrs;
   });
 
   this.get('loan-storage/loan-history', ({ loanactions }, request) => {
@@ -301,6 +300,10 @@ export default function config() {
       loansHistory: [],
       totalRecords: 0,
     };
+  });
+
+  this.put('/loan-storage/loans/:id', (_, request) => {
+    return JSON.parse(request.requestBody);
   });
 
   this.get('/circulation/requests', function ({ requests }) {
@@ -926,37 +929,5 @@ export default function config() {
       ],
       amount : 400.00
     };
-  });
-
-  this.post('/feefine-reports/refund', {
-    reportData: [
-      {
-        patronName: 'Ellis, James S',
-        patronBarcode: '12345',
-        patronId: '54d9cdf3-c928-4919-8d73-fc63c2155765',
-        patronGroup: 'Staff',
-        feeFineType: 'Lost item processing fee',
-        dateBilled: '12/4/2020, 1:13 PM',
-        billedAmount: '100.00',
-        paidAmount: '100.00',
-        paymentMethod: 'Cash',
-        transactionInfo: 'ECON #5820',
-        transferredAmount: '50.00',
-        transferAccount: 'Account',
-        feeFineId: '17d9cdf3-c928-4919-8d73-fc63c2155764',
-        refundDate: '12/4/2020, 1:13 PM',
-        refundAmount: '100.00',
-        refundAction: 'Refunded fully',
-        refundReason: 'Overpaid',
-        staffInfo: 'Refund staff info',
-        patronInfo: 'Refund patron info',
-        itemBarcode: '123456789',
-        instance: 'It\'s raining cats and dogs',
-        actionCompletionDate: '12/4/2020, 1:13 PM',
-        staffMemberName: 'John Smith',
-        actionTaken: '12/4/2020, 1:13 PM'
-      }
-    ],
-    totalRecords: 1,
   });
 }