+++ test/bigtest/network/factories/feefineaction.js
@@ -10,7 +10,7 @@ export default Factory.extend({
   createdAt: 'Main Circ',
   dateAction: '2019-04-22T22:07:07.343+0000',
   source: 'User, Test',
-  transactionInformation: '-',
+  transactionInformation: '',
   typeAction: 'Late',
   userId: '2da52ef1-81b8-4d7d-9b56-aba7f6885a72'
 });